reworked python script
first python function!

This python script works the same way as the powershell script of the same name

it also appears there is no -- or ++ in python.

The other thing of note for me here currently is putting variables into strings as i do in the print command on line 21 of the python script committed. % (var1,var2...,varX) and %s/%d are the expected var type to be entered